bot/requirements.txt
# Telegram
python-telegram-bot==21.6

# LangChain stack
langchain==1.2.10
langchain-openai==1.1.10
langchain-community==0.4.1
#langchain-core==0.3.15
langchain-core==1.2.17
pydantic-settings==2.13.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
psycopg[binary]

# Config
#pydantic-settings==2.6.1
python-dotenv==1.0.1
